treat binding values as expressions
